+++ src/web/system.py
@@ -46,7 +46,7 @@ from ombrebrain.policy import RedLineContract, RedLineFeatureSpec, SurfaceDecisi
 from ombrebrain.protocol import PublicToolDesignContract, PublicToolSpec
 from ombrebrain.resilience import CrashRecoveryContract, CrashRecoveryPlan, PathStep
 from ombrebrain.retrieval import SurfaceContextCompiler
-from deployment_profile import effective_configuration_report
+from ombrebrain.security.deployment_profile import effective_configuration_report
 from utils import config_file_path
 
 try:
@@ -59,10 +59,7 @@ try:
 except ImportError:  # pragma: no cover
     from ..utils import parse_bool  # type: ignore
 
-try:
-    from vault_health import inspect_vault  # type: ignore
-except ImportError:  # pragma: no cover
-    from ..vault_health import inspect_vault  # type: ignore
+from ombrebrain.storage.vault_health import inspect_vault
 
 _LOGS_DEFAULT_LIMIT = 200
 _LOGS_MAX_LIMIT = 2000